add where_by_date to construct granular time range in query
